requirements.txt
requests==2.31.0
python-dotenv==1.0.0
discord.py==2.3.2
aiohttp==3.9.1
markitdown[pptx]==0.0.1a2
defusedxml==0.7.1
